requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
asttokens==3.0.0
beautifulsoup4==4.13.3
bella==2.1.6
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.5.14
decorator==5.1.1
Events==0.5
executing==2.2.0
fire==0.3.1
idna==3.10
ipython==8.32.0
isodate==0.6.1
jedi==0.19.2
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pip_search==0.0.13
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-youtube==0.6.4
pytube3==9.6.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
termcolor==2.5.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
wcwidth==0.2.13
wxPython==4.2.2
yt-dlp==2025.1.26
ytdl==1.0.0rc5
